BB-7

+++ lib/src/table_feature/table_app_bar.dart
@@ -3,6 +3,7 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:get/get.dart';
 import 'package:numberpicker/numberpicker.dart';
 
+import '../auth/auth_service.dart';
 import '../management/table_controller.dart';
 
 class TableAppBar extends AppBar {
@@ -20,37 +21,38 @@ class _TableAppBarState extends State<TableAppBar> {
     return AppBar(
       title: const Text("Tischplan"),
       actions: [
-        IconButton(
-          icon: const Icon(FontAwesomeIcons.plusMinus),
-          onPressed: () => Get.defaultDialog(
-            titlePadding: const EdgeInsets.symmetric(vertical: 22.0),
-            title: "Tische anpassen",
-            content: StatefulBuilder(
-              builder: (_, setState) => NumberPicker(
-                minValue: 0,
-                maxValue: 50,
-                value: _pickerValue,
-                onChanged: (value) => setState(() => _pickerValue = value),
+        if (AuthService.to.isAdmin.value)
+          IconButton(
+            icon: const Icon(FontAwesomeIcons.plusMinus),
+            onPressed: () => Get.defaultDialog(
+              titlePadding: const EdgeInsets.symmetric(vertical: 22.0),
+              title: "Tische anpassen",
+              content: StatefulBuilder(
+                builder: (_, setState) => NumberPicker(
+                  minValue: 0,
+                  maxValue: 50,
+                  value: _pickerValue,
+                  onChanged: (value) => setState(() => _pickerValue = value),
+                ),
               ),
+              actions: [
+                TextButton(
+                  onPressed: () async {
+                    await TableController.to.addNTables(_pickerValue);
+                    Get.back();
+                  },
+                  child: const Text("Hinzufügen"),
+                ),
+                TextButton(
+                  onPressed: () async {
+                    await TableController.to.deleteNTables(_pickerValue);
+                    Get.back();
+                  },
+                  child: const Text("Löschen"),
+                ),
+              ],
             ),
-            actions: [
-              TextButton(
-                onPressed: () async {
-                  await TableController.to.addNTables(_pickerValue);
-                  Get.back();
-                },
-                child: const Text("Hinzufügen"),
-              ),
-              TextButton(
-                onPressed: () async {
-                  await TableController.to.deleteNTables(_pickerValue);
-                  Get.back();
-                },
-                child: const Text("Löschen"),
-              ),
-            ],
-          ),
-        )
+          )
       ],
     );
   }
